updated test_predictions float64 to float32

+++ tests/test_predictions.py
@@ -38,7 +38,7 @@ def test_make_prediction(sample_input_data):
     # Then
     predictions = result.get("predictions")
     assert isinstance(predictions, np.ndarray)
-    assert isinstance(predictions[0], np.float64)
+    assert isinstance(predictions[0], np.float32)
     assert result.get("errors") is None
     assert len(predictions) == expected_no_predictions
